pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "transfermarket"
version = "0.0.0"
authors = [
  { name="Omar Crosby", email="omar.crosby@gmail.com" },
]
description = "A python module for retrieving information from https://www.transfermarkt.com."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License (GPL)",
    "Operating System :: OS Independent"
]
keywords = [
    "transfermarkt",
    "football",
    "soccer",
    "football club",
    "football player",
    "football manager",
    "football league",
    "football competition",
    "football transfer",
    "football statistics",
    "football data"
]

[project.urls]
"Homepage" = "https://github.com/ocrosby/transfermarkt"
"Bug Tracker" = "https://github.com/ocrosby/transfermarkt/issues"
"Target URL" = "https://www.transfermarkt.com"

[tool.poetry]
name = "transfermarket"
version = "0.0.0"
description = "A python module for retrieving information from https://www.transfermarkt.com."
authors = ["Omar Crosby <omar.crosby@gmail.com>"]

[tool.pytest.ini_options]
addopts = "-ra -q --cov-report=term-missing --cov-report=html"
testpaths = [
    "tests/unit"
]

[tool.coverage.run]
omit = [
    "tests/*",
]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.9.3"
requests = "^2.25.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^22.10.0"
pylint = "^2.11.1"
pytest = "^7.1.3"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
coverage = "^6.4.4"
flake8 = "^5.0.4"
radon = "^5.1.0"
isort = "^5.5.0"
invoke = "^1.7.3"
mypy = "^0.991"
requests-mock = "^1.9.3"
Flake8-pyproject = "^1.1.0.post0"
freezegun = "^1.0.0"
python-semantic-release = "^7.19.0"

[tool.flake8]
ignore = ['C901', 'E203', 'F401', 'W503']
max-line-length = 88
exclude = ['.env', '.venv', '.env', 'env', 'migrations', '__pycache__', 'manage.py', 'settings.py']

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "C0103", # Invalid name
    "C0301", # Line too long
    "C0114", # Missing module docstring
    "C0115", # Missing class docstring
    "C0116", # Missing function or method docstring
    "R0902", # Too many instance attributes
    "R0903", # Too few public methods
    "W0212", # Access to a protected member
    "W0613", # Unused argument
    "W0621", # Redefining name from outer scope
    "W1514" # Using open without explicitly specifying an encoding
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 119
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.semantic_release]
version_source = "commit"
version_variable = [
    "src/transfermarket/__init__.py:__version__",
    "pyproject.toml:version"
]
version_toml = [
    "pyproject.toml:project.version",
    "pyproject.toml:tool.poetry.version"
]
version_pattern = [
    "README.md:rev: v{version}"
]
major_on_zero = true
branch = "main"
upload_to_pypi = false
upload_to_release = true
commit_subject = "Updating version from {current_version} to {new_version}"
commit_message = "chore(release): {current_version} → {new_version}"
build_command = "pip install --upgrade pip && pip install poetry && poetry build"
